+++ just_audio_platform_interface/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.2.1
 
 - Added support for HLS title and artist metadata.
+- Update minimum flutter version to 3.0.
 
 ## 4.2.0